Add votes to deck preview

+++ postgres/init.sql
@@ -215,7 +215,7 @@ CREATE VIEW mythgard.deck_preview as
          deck.created as deck_created,
          array_agg(DISTINCT faction.name) as factions,
          sum(essence_costs.essence)::int as essence_cost,
-         count(deck_vote) as votes
+         count(deck_vote)::int as votes
   FROM mythgard.deck
   JOIN mythgard.card_deck
     ON card_deck.deck_id = deck.id
